Refactor role assignment logic for AI Foundry and update project ARM path resolution
